requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.12
django-appconf==1.0.5
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-filter==21.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-redis==5.2.0
django-select2==7.10.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
protobuf==3.19.4
psycopg2-binary==2.9.3
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytimeparse==1.1.8
pytz==2022.1
PyYAML==6.0
redis==4.2.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tabulate==0.8.9
tqdm==4.63.1
typing_extensions==4.1.1
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
